Fix css loading issues
In some cases, css were not loading. This was probably caused by some kind of race between the loading of the page and the loading of the user script or user script manager: the mutation observer would only be added AFTER the css were loaded. These changes fix that.
Additionnally, I put the functions outside of the big one (for aesthetic reasons, I guess...).